refactor(Apostille, ApostilleAccount): Moved some code from Apostille class to ApostilleAccount clas
We refactor some methods (Create, Update, Announce, etc) from Apostille class, so ApostilleAccount
can use the method as well
