+++ src/IReadyTransaction.ts
